requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncio==3.4.3
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0 